requirements.txt
add-trailing-comma
autopep8
black==21.5b1
celery==5.3.6
cryptography==43.0.0
dill==0.3.7
Django==4.2
django-admin-rangefilter
django-axes==6.2.0
django-bleach==3.1.0
django-celery-beat==2.5.0
django-cors-headers
django-extensions==3.2.3
django-filter==23.3
django-import-export==3.3.3
django-model-utils==4.3.1
django-multiselectfield==0.1.12 # This package needs to be removed
django-redis==5.4.0
django-ses
django-storages
django-tools==0.54.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.3.1
drf-dynamic-read==0.1.1
drf-extra-fields==3.7.0
drf-yasg
flake8
googlemaps
ijson

ipython==8.26.0
mapbox
newrelic
pandas==2.1.3
pdfkit==1.0.0
pep8
Pillow==10.1.0
pre-commit
psycopg2-binary==2.9.9
pycodestyle
pycryptodome==3.15.0
pygments==2.17.2
redis==5.0.1
PyPDF2
requests
sentry-sdk
shapely
ujson
